refactor(rpc): move transactions + epoch handlers out of routes/mod.rs (backlog #12 phase 2f + 2g)
The last two slices of the routes refactor. Transactions (send, get,
mempool) land in routes/transactions.rs with the Transaction import
following them. Epoch (current, history) land in routes/epoch.rs.

After this mod.rs is ~420 lines of router wiring plus the shared
api_err helper, down from two-thousand-plus at the start of the refactor.
Phase 2 of backlog #12 is done.
